// app/(root)/blog/types.ts
export interface Post {
    id: number;
    date: string;
    modified: string;
    slug: string;
    status: string;
    type: string;
    title: {
      rendered: string;
    };
    content: {
      rendered: string;
    };
    excerpt: {
      rendered: string;
    };
    featured_media: number;
    media_url?: string;
  }
  
  export interface Category {
    id: number;
    count: number;
    description: string;
    link: string;
    name: string;
    slug: string;
    taxonomy: string;
    parent: number;
  }
  